Allowing pore scale models to access propname string in paramlist
We don't allow pore-scale models to have any arg with propname string in it since it probably would cause an infinite cycle, i.e. model A that writes to 'pore.blah' can't have 'pore.blah' as another argument. But in Multiphase object we need this since 'pore.blah' in the Multiphase object is different than 'pore.blah' of the constituent phases.